Bug 1994937 - Change argument of WebRenderAPI::UpdateDebugFlags() to 64 bit r=gfx-reviewers,nical
Bug 1946361 changed DebugFlags from 32 bit to 64 bit. But it forgot to change the argument of WebRenderAPI::UpdateDebugFlags() to 64 bit.

Differential Revision: https://phabricator.services.mozilla.com/D269006

+++ gfx/webrender_bindings/WebRenderAPI.cpp
@@ -488,7 +488,7 @@ wr::WebRenderAPI* WebRenderAPI::GetRootAPI() {
   return this;
 }
 
-void WebRenderAPI::UpdateDebugFlags(uint32_t aFlags) {
+void WebRenderAPI::UpdateDebugFlags(uint64_t aFlags) {
   wr_api_set_debug_flags(mDocHandle, wr::DebugFlags{aFlags});
 }
